extension changes

+++ extension/opencart/catalog/model/shipping/free.js
@@ -14,7 +14,7 @@ class Free extends \Opencart\System\Engine\Model {
 	function getQuote(array $address): array {
 		this.load.language('extension/opencart/shipping/free');
 
-		$query = this.db.query("SELECT * FROM `" . DB_PREFIX . "zone_to_geo_zone` WHERE `geo_zone_id` = '" . (int)this.config.get('shipping_free_geo_zone_id') . "' AND `country_id` = '" . (int)$address['country_id'] . "' AND (`zone_id` = '" . (int)$address['zone_id'] . "' OR `zone_id` = '0')");
+		$query = this.db.query("SELECT * FROM `" . DB_PREFIX . "zone_to_geo_zone` WHERE `geo_zone_id` = '" . this.config.get('shipping_free_geo_zone_id') . "' AND `country_id` = '" . $address['country_id'] . "' AND (`zone_id` = '" . $address['zone_id'] . "' OR `zone_id` = '0')");
 
 		if (!this.config.get('shipping_free_geo_zone_id')) {
 			$status = true;
